Added Warranty Repository

+++ src/main/java/com/app/warrantychecker/model/Warranty.java
@@ -15,7 +15,7 @@ public class Warranty {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int warrantyID;
+    private long warrantyID;
 
     @CreatedDate
     private Date createdDate;
